move files away from ng naming scheme
