+++ SpotBot/src/app.ts
@@ -2,12 +2,12 @@ import { Client, GuildMember, PartialGuildMember, Role, GatewayIntentBits, Event
 require('dotenv').config()
 import 'reflect-metadata';
 import { container } from 'tsyringe';
-import { AppService, constructLeaveMessage } from './app.service';
-import { LogService } from './services/log.service';
-import { FileService } from './services/file.service';
-import { ConfigurationService } from './services/configuration/configuration.service';
-import { MessageService } from './services/message/message.service';
-import { WelcomeService } from './services/welcome/welcome.service';
+import { AppService, constructLeaveMessage } from './app_service';
+import { LogService } from './services/log_service';
+import { FileService } from './services/file_service';
+import { ConfigurationService } from './services/configuration/configuration_service';
+import { MessageService } from './services/message/message_service';
+import { WelcomeService } from './services/welcome/welcome_service';
 import { PingCommand, ServerCommand, SetWelcomeCommand, ReactionCommand } from './appCommands';
 import { ICommandServices } from './interfaces/ICommandServices';
