+++ SpotBot/src/interfaces/ICommandServices.ts
@@ -1,6 +1,6 @@
-import { FileService } from "../services/file.service";
-import { LogService } from "../services/log.service";
-import { WelcomeService } from "../services/welcome/welcome.service";
+import { FileService } from "../services/file_service";
+import { LogService } from "../services/log_service";
+import { WelcomeService } from "../services/welcome/welcome_service";
 
 export interface ICommandServices {
     logService: LogService,
