+++ SpotBot/src/services/log_service.ts
